# Replace C*Pos intermediate variables with expressions