--- a/drake/multibody/dev/test/global_inverse_kinematics_reachable_test.cc
+++ b/drake/multibody/dev/test/global_inverse_kinematics_reachable_test.cc
@@ -32,7 +32,7 @@ TEST_F(KukaTest, ReachableTest) {
     EXPECT_EQ(sol_result, SolutionResult::kSolutionFound);
 
     double pos_tol = 0.06;
-    double orient_tol = 0.2;
+    double orient_tol = 0.22;
     CheckGlobalIKSolution(pos_tol, orient_tol);
     // Now call nonlinear IK with the solution from global IK as the initial
     // seed. If the global IK provides a good initial seed, then the nonlinear