--- a/drake/solvers/rotation_constraint.h
+++ b/drake/solvers/rotation_constraint.h
@@ -5,6 +5,7 @@
 #include <utility>
 #include <vector>
 
+#include "drake/common/symbolic_expression.h"
 #include "drake/solvers/decision_variable.h"
 #include "drake/solvers/mathematical_program.h"
 
@@ -91,8 +92,8 @@ void AddRotationMatrixOrthonormalSocpConstraint(
     const Eigen::Ref<const MatrixDecisionVariable<3, 3>>& R);
 
 using AddRotationMatrixMcCormickEnvelopeReturnType =
-std::tuple<std::vector<MatrixDecisionVariable<3, 3>>,
-           std::vector<MatrixDecisionVariable<3, 3>>,
+std::tuple<std::vector<Eigen::Matrix<drake::symbolic::Expression, 3, 3>>,
+           std::vector<Eigen::Matrix<drake::symbolic::Expression, 3, 3>>,
            std::vector<MatrixDecisionVariable<3, 3>>,
            std::vector<MatrixDecisionVariable<3, 3>>>;
 /**