This replaces some continuous variables that were carried around to make writing the rotation constraints easier. Instead, they are now formed as expressions, simplifying the optimization.

This has been deployed on my stuff for a while, but the timing differences it makes on tests relying on the rotations show that it does increase setup time, but makes longer-running solves more efficient:

All times reported below are found by running `time <the test binary>` and taking the user time (which I interpret as total CPU time spent for the process running non-syscall code). I tested each case 3 times, and numbers didn't change more than a few percent. The first two cases have lots of short-running solves, and last two cases have much longer-running solves.

| Test | Total CPU core time without change   | Total CPU core time with change |
| -------------- | ------------- | ------------- |
| rotation_constraint_test (all) | 1m 7s  | 1m 24s  |
| TestMcCormick | 0m 1.4s  | 0m 2.0s  |
| rotation_constraint_limit_test (all) | 37m 0s | 29-33m 0s |
| global_inverse_kinematics_collision_avoidance_test | 10m 40s | 4m 10s |

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/robotlocomotion/drake/6533)
<!-- Reviewable:end -->