@hongkai-dai 2017-07-07
+@hongkai-dai  for feature review

---

Review status: 0 of 3 files reviewed at latest revision, all discussions resolved.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/robotlocomotion/drake/6533#-:-KoTa4on5-AGpOoezZcz:b-8zd4a7)*
<!-- Sent from Reviewable.io -->