@hongkai-dai 2017-07-07
Review status: 0 of 3 files reviewed at latest revision, 2 unresolved discussions, some commit checks failed.

---

*[drake/solvers/rotation_constraint.cc, line 921 at r1](https://reviewable.io:443/reviews/robotlocomotion/drake/6533#-KoTbqlxw4juD8xvdntO:-KoTbqlxw4juD8xvdntP:b-3tcgqa) ([raw file](https://github.com/robotlocomotion/drake/blob/1a9c71dc67eadc8abbe118df50384304663ca922/drake/solvers/rotation_constraint.cc#L921)):*
> ```cpp
>   // pass into the constraints changes for the k=N-1 case.  Otherwise we
>   // could do a simple substitution everywhere.
>   // TODO(russt): Use symbolic constraints and remove these decision variables!
> ```

Remove this TODO?

---

*[drake/solvers/rotation_constraint.cc, line 929 at r1](https://reviewable.io:443/reviews/robotlocomotion/drake/6533#-KoTcA0UqV0kW4x5fesA:-KoTcA0UqV0kW4x5fesB:b-f9gtl) ([raw file](https://github.com/robotlocomotion/drake/blob/1a9c71dc67eadc8abbe118df50384304663ca922/drake/solvers/rotation_constraint.cc#L929)):*
> ```cpp
>     BRneg.push_back(
>         prog->NewBinaryVariables<3, 3>("BRneg" + std::to_string(k)));
>     CRpos.push_back(Eigen::Matrix3d::Ones());
> ```

Why we need to push back ones here? I think we can just do 
```C++
CRpos.reserve(num_binary_vars_per_half_axis);
CRneg.reserve(num_binary_vars_per_half_axis);
for (int k = 0; k < num_binary_vars_per_half_axis; k++) {
    BRpos.push_back(
        prog->NewBinaryVariables<3, 3>("BRpos" + std::to_string(k)));
    BRneg.push_back(
        prog->NewBinaryVariables<3, 3>("BRneg" + std::to_string(k)));
}
for (int k = 0; k < num_binary_vars_per_half_axis - 1; k++) {
  CRpos.push_back(BRpos[k] - BRpos[k + 1]);
  CRneg.push_back(BRneg[k] - BRneg[k + 1]);
}
CRpos.push_back(BRpos[num_binary_vars_per_half_axis - 1].cast<symbolic::Expression>());
CRneg.push_back(BRneg[num_binary_vars_per_half_axis - 1].cast<symbolic::Expression>());
```
And remove the changes on `CRpos, CRneg` down below.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/robotlocomotion/drake/6533)*
<!-- Sent from Reviewable.io -->