@hongkai-dai 2017-07-07
@drake-jenkins-bot mac-clang-bazel-experimental please

---

Review status: 0 of 3 files reviewed at latest revision, 2 unresolved discussions, some commit checks failed.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/robotlocomotion/drake/6533#-:-KoUHmT9l1Kl2yZDXkyO:bjwebpu)*
<!-- Sent from Reviewable.io -->