@hongkai-dai 2017-07-08
Reviewed 2 of 3 files at r1.
Review status: 2 of 3 files reviewed at latest revision, 8 unresolved discussions, some commit checks failed.

---

*[drake/solvers/rotation_constraint.cc, line 569 at r1](https://reviewable.io:443/reviews/robotlocomotion/drake/6533#-KoUY3Q1DYncgNn0NTFB:-KoUY3Q1DYncgNn0NTFC:b-o7mjrv) ([raw file](https://github.com/robotlocomotion/drake/blob/1a9c71dc67eadc8abbe118df50384304663ca922/drake/solvers/rotation_constraint.cc#L569)):*
> ```cpp
>     MathematicalProgram* prog, const VectorDecisionVariable<3>& v,
>     const std::vector<Eigen::Matrix<Expression, 3, 1>>& cpos,
>     const std::vector<Eigen::Matrix<Expression, 3, 1>>& cneg,
> ```

BTW, we have `Vector3<Expression>`, if we include `drake/common/eigen_types.h`.

---

*[drake/solvers/rotation_constraint.cc, line 585 at r1](https://reviewable.io:443/reviews/robotlocomotion/drake/6533#-KoUXwgoL7QZY5XLfgVY:-KoUXwgoL7QZY5XLfgVZ:b-7ropln) ([raw file](https://github.com/robotlocomotion/drake/blob/1a9c71dc67eadc8abbe118df50384304663ca922/drake/solvers/rotation_constraint.cc#L585)):*
> ```cpp
>         box_max(2) = EnvelopeMinValue(zi + 1, N);
> 
>         Eigen::Matrix<Expression, 3, 1> this_cpos, this_cneg;
> ```

ditto

---

*[drake/solvers/rotation_constraint.cc, line 635 at r1](https://reviewable.io:443/reviews/robotlocomotion/drake/6533#-KoUYIbXRsPb4IzrtLFj:-KoUYIbXRsPb4IzrtLFk:b-7ropln) ([raw file](https://github.com/robotlocomotion/drake/blob/1a9c71dc67eadc8abbe118df50384304663ca922/drake/solvers/rotation_constraint.cc#L635)):*
> ```cpp
>             }
>             Eigen::Vector3d orthant_u;
>             Eigen::Matrix<Expression, 3, 1> orthant_c;
> ```

ditto

---

*[drake/solvers/rotation_constraint.cc, line 691 at r1](https://reviewable.io:443/reviews/robotlocomotion/drake/6533#-KoUYHabEsshn2Y3NYn-:-KoUYHabEsshn2Y3NYn0:b-7ropln) ([raw file](https://github.com/robotlocomotion/drake/blob/1a9c71dc67eadc8abbe118df50384304663ca922/drake/solvers/rotation_constraint.cc#L691)):*
> ```cpp
> 
>             Eigen::Vector3d orthant_normal;
>             Eigen::Matrix<Expression, 3, 1> orthant_c;
> ```

ditto

---

*[drake/solvers/rotation_constraint.cc, line 759 at r1](https://reviewable.io:443/reviews/robotlocomotion/drake/6533#-KoUYgnXRA0g-s4UutNd:-KoUYgnXRA0g-s4UutNe:btm8kaw) ([raw file](https://github.com/robotlocomotion/drake/blob/1a9c71dc67eadc8abbe118df50384304663ca922/drake/solvers/rotation_constraint.cc#L759)):*
> ```cpp
>               //   we can impose a tighter Lorentz cone constraint
>               //     [|sin(theta)|, n'*v1, n'*v2] is in the Lorentz cone.
>               a << orthant_normal, -2, -2, -2;
> ```

`a` is not used anymore?

If setup time becomes a concern, we will avoid this symbolic constraint. We do need to parse the same symbolic expression for twice here, one for >=, and one for <= .

---

*[drake/solvers/rotation_constraint.h, line 95 at r1](https://reviewable.io:443/reviews/robotlocomotion/drake/6533#-KoUZ9ObMasOlZo3TCCo:-KoUZ9ObMasOlZo3TCCp:b-6390mj) ([raw file](https://github.com/robotlocomotion/drake/blob/1a9c71dc67eadc8abbe118df50384304663ca922/drake/solvers/rotation_constraint.h#L95)):*
> ```objc
> 
> using AddRotationMatrixMcCormickEnvelopeReturnType =
> std::tuple<std::vector<Eigen::Matrix<drake::symbolic::Expression, 3, 3>>,
> ```

BTW, we also have `Matrix3<symbolic::Expression>`.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/robotlocomotion/drake/6533)*
<!-- Sent from Reviewable.io -->